requirements.txt
psycopg2~=2.8.5
Jinja2~=2.11.2
Flask~=1.1.2
PyYAML~=5.3.1
sshtunnel~=0.1.5
python-telegram-bot~=12.8
lxml~=4.5.1
